pyproject.toml
[tool.poetry]
name = "backoff-test"
version = "0.0.1"
description = "Reproduce backoff error in Otel grpc exporter"
authors = ["pete.topleft@gmail.com"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.0"
opentelemetry-sdk = "^1.13.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.13.0"
uvicorn = {extras = ["standard"], version = "^0.18.3"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
